requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
build==1.0.3
certifi>=2024.07.04
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
fonttools==4.53.1
gravis==0.1.0
greenlet==3.0.3
idna>=3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.9.1
networkx==3.3
numpy==2.0.1
packaging==23.2
pandoc==2.3
pillow==10.4.0
plumbum==1.8.2
ply==3.11
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
requests==2.31.0
scipy==1.14.0
setuptools>=70.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3>=2.2.2
werkzeug>=3.0.6
wheel==0.42.0
google-auth
google-auth-oauthlib
google-api-python-client